fix meshgrid issue for AP metric with distance 1

+++ keypoint_detection/keypoint_detection/src/keypoint_utils.py
@@ -7,7 +7,7 @@ from skimage.feature import peak_local_max
 
 
 def gaussian_heatmap(
-    image_size: Tuple[int, int], center: Union[Tuple[int, int], List[int]], sigma: torch.Tensor
+    image_size: Tuple[int, int], center: Union[Tuple[float, float], List[float]], sigma: torch.Tensor
 ) -> torch.Tensor:
     """
     Creates a Gaussian blob heatmap for a single keypoint.
@@ -22,18 +22,20 @@ def gaussian_heatmap(
         Torch.Tensor: A tensor with zero background, specified size and a Gaussian heatmap around the center.
     """
 
-    u_axis = torch.linspace(0, image_size[1] - 1, image_size[1]) - center[0]
-    v_axis = torch.linspace(0, image_size[0] - 1, image_size[0]) - center[1]
+    # cast keypoints (center) to ints to make grid align with pixel raster.
+    #  Otherwise, the AP metric for  d = 1 will not result in 1
+    #  if the gt_heatmaps are used as input.
+    u_axis = torch.linspace(0, image_size[1] - 1, image_size[1]) - int(center[0])
+    v_axis = torch.linspace(0, image_size[0] - 1, image_size[0]) - int(center[1])
     # create grid values in 2D with x and y coordinate centered aroud the keypoint
     xx, yy = torch.meshgrid(v_axis, u_axis)
-
     ## create gaussian around the centered 2D grids $ exp ( -0.5 (x**2 + y**2) / sigma**2)$
     heatmap = torch.exp(-0.5 * (torch.square(xx) + torch.square(yy)) / torch.square(sigma))
     return heatmap
 
 
 def generate_keypoints_heatmap(
-    image_size: Tuple[int, int], keypoints: List[Tuple[int, int]], sigma: float
+    image_size: Tuple[int, int], keypoints: List[Tuple[float, float]], sigma: float
 ) -> torch.Tensor:
     """
     Generates heatmap with gaussian blobs for each keypoint, using the given sigma.
